Remove less-useful base classes
Make GLSLMaterial inherit from BaseMaterial
