--- engine/src/main/java/org/terasology/rendering/base/BaseMesh.java
@@ -1,63 +1,0 @@
-/*
- * Copyright 2014 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.terasology.rendering.base;
-
-import gnu.trove.list.TFloatList;
-
-import org.terasology.asset.AbstractAsset;
-import org.terasology.asset.AssetUri;
-import org.terasology.math.AABB;
-import org.terasology.rendering.assets.mesh.Mesh;
-import org.terasology.rendering.assets.mesh.MeshData;
-
-public abstract class BaseMesh extends AbstractAsset<MeshData> implements Mesh {
-
-    protected MeshData data;
-    protected AABB aabb;
-
-    public BaseMesh(AssetUri uri, MeshData data) {
-        super(uri);
-        reload(data);
-    }
-
-    public abstract void render();
-
-    @Override
-    public void reload(MeshData data) {
-        this.data = data;
-        this.aabb = AABB.createEncompasing(data.getVertices());
-    }
-
-    @Override
-    public void dispose() {
-        data = null;
-    }
-
-    @Override
-    public boolean isDisposed() {
-        return data == null;
-    }
-
-    @Override
-    public AABB getAABB() {
-        return aabb;
-    }
-
-    @Override
-    public TFloatList getVertices() {
-        return data.getVertices();
-    }
-}
