--- engine/src/main/java/org/terasology/rendering/base/BaseTexture.java
@@ -1,90 +1,0 @@
-/*
- * Copyright 2014 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.terasology.rendering.base;
-
-import org.terasology.asset.AbstractAsset;
-import org.terasology.asset.AssetUri;
-import org.terasology.math.Rect2f;
-import org.terasology.math.Vector2i;
-import org.terasology.rendering.assets.texture.Texture;
-import org.terasology.rendering.assets.texture.TextureData;
-
-public abstract class BaseTexture extends AbstractAsset<TextureData> implements Texture {
-
-    protected TextureData textureData;
-
-    public BaseTexture(AssetUri uri, TextureData data) {
-        super(uri);
-        reload(data);
-    }
-
-    public abstract int getId();
-
-    @Override
-    public void reload(TextureData textureData) {
-        this.textureData = textureData;
-    }
-
-    @Override
-    public void dispose() {
-        this.textureData = null;
-    }
-
-    @Override
-    public boolean isDisposed() {
-        return textureData == null;
-    }
-
-    @Override
-    public Texture getTexture() {
-        return this;
-    }
-
-    @Override
-    public Rect2f getRegion() {
-        return FULL_TEXTURE_REGION;
-    }
-
-    @Override
-    public int getWidth() {
-        return textureData.getWidth();
-    }
-
-    @Override
-    public int getHeight() {
-        return textureData.getHeight();
-    }
-
-    @Override
-    public Vector2i size() {
-        return new Vector2i(getWidth(), getHeight());
-    }
-
-    @Override
-    public WrapMode getWrapMode() {
-        return textureData.getWrapMode();
-    }
-
-    @Override
-    public FilterMode getFilterMode() {
-        return textureData.getFilterMode();
-    }
-
-    @Override
-    public TextureData getData() {
-        return textureData;
-    }
-}
