+++ engine/src/main/java/org/terasology/rendering/headless/HeadlessMesh.java
@@ -15,14 +15,48 @@
  */
 package org.terasology.rendering.headless;
 
-import org.terasology.asset.AssetUri;
-import org.terasology.rendering.assets.mesh.MeshData;
-import org.terasology.rendering.base.BaseMesh;
+import gnu.trove.list.TFloatList;
 
-public class HeadlessMesh extends BaseMesh {
+import org.terasology.asset.AbstractAsset;
+import org.terasology.asset.AssetUri;
+import org.terasology.math.AABB;
+import org.terasology.rendering.assets.mesh.Mesh;
+import org.terasology.rendering.assets.mesh.MeshData;
+
+public class HeadlessMesh extends AbstractAsset<MeshData> implements Mesh {
+
+    protected MeshData data;
+    protected AABB aabb;
 
     public HeadlessMesh(AssetUri uri, MeshData data) {
-        super(uri, data);
+        super(uri);
+        reload(data);
+    }
+
+    @Override
+    public void reload(MeshData data) {
+        this.data = data;
+        this.aabb = AABB.createEncompasing(data.getVertices());
+    }
+
+    @Override
+    public void dispose() {
+        data = null;
+    }
+
+    @Override
+    public boolean isDisposed() {
+        return data == null;
+    }
+
+    @Override
+    public AABB getAABB() {
+        return aabb;
+    }
+
+    @Override
+    public TFloatList getVertices() {
+        return data.getVertices();
     }
 
     @Override
