+++ engine/src/main/java/org/terasology/rendering/opengl/GLSLMaterial.java
@@ -16,8 +16,6 @@
 
 package org.terasology.rendering.opengl;
 
-import com.google.common.base.Objects;
-import com.google.common.collect.Sets;
 import gnu.trove.iterator.TIntIntIterator;
 import gnu.trove.map.TIntIntMap;
 import gnu.trove.map.TIntObjectMap;
@@ -25,29 +23,7 @@ import gnu.trove.map.TObjectIntMap;
 import gnu.trove.map.hash.TIntIntHashMap;
 import gnu.trove.map.hash.TIntObjectHashMap;
 import gnu.trove.map.hash.TObjectIntHashMap;
-import org.lwjgl.opengl.GL13;
-import org.lwjgl.opengl.GL20;
-import org.lwjgl.opengl.Util;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.terasology.asset.AbstractAsset;
-import org.terasology.asset.AssetUri;
-import org.terasology.registry.CoreRegistry;
-import org.terasology.math.MatrixUtils;
-import org.terasology.rendering.ShaderManager;
-import org.terasology.rendering.assets.material.Material;
-import org.terasology.rendering.assets.material.MaterialData;
-import org.terasology.rendering.assets.shader.ShaderParameterMetadata;
-import org.terasology.rendering.assets.shader.ShaderProgramFeature;
-import org.terasology.rendering.assets.texture.Texture;
-import org.terasology.rendering.cameras.Camera;
-import org.terasology.rendering.shader.ShaderParameters;
 
-import javax.vecmath.Matrix3f;
-import javax.vecmath.Matrix4f;
-import javax.vecmath.Tuple2f;
-import javax.vecmath.Tuple3f;
-import javax.vecmath.Tuple4f;
 import java.nio.FloatBuffer;
 import java.util.Arrays;
 import java.util.Collections;
@@ -55,11 +31,32 @@ import java.util.EnumSet;
 import java.util.Map;
 import java.util.Set;
 
+import javax.vecmath.Matrix3f;
+import javax.vecmath.Matrix4f;
+
+import org.lwjgl.opengl.GL13;
+import org.lwjgl.opengl.GL20;
+import org.lwjgl.opengl.Util;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.terasology.asset.AssetUri;
+import org.terasology.math.MatrixUtils;
+import org.terasology.rendering.ShaderManager;
+import org.terasology.rendering.assets.material.MaterialData;
+import org.terasology.rendering.assets.shader.ShaderParameterMetadata;
+import org.terasology.rendering.assets.shader.ShaderProgramFeature;
+import org.terasology.rendering.assets.texture.Texture;
+import org.terasology.rendering.base.BaseMaterial;
+import org.terasology.rendering.shader.ShaderParameters;
+
+import com.google.common.base.Objects;
+import com.google.common.collect.Sets;
+
 /**
  * @author Immortius
  * @author Benjamin Glatzel <benjamin.glatzel@me.com>
  */
-public class GLSLMaterial extends AbstractAsset<MaterialData> implements Material {
+public class GLSLMaterial extends BaseMaterial {
 
     private static final Logger logger = LoggerFactory.getLogger(GLSLMaterial.class);
 
@@ -79,8 +76,7 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
     private ShaderParameters shaderParameters;
 
     public GLSLMaterial(AssetUri uri, MaterialData data) {
-        super(uri);
-        shaderManager = CoreRegistry.get(ShaderManager.class);
+        super(uri, data);
         reload(data);
     }
 
@@ -248,21 +244,8 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
     }
 
     @Override
-    public void setCamera(Camera camera) {
-        setMatrix4("viewMatrix", camera.getViewMatrix());
-        setMatrix4("projMatrix", camera.getProjectionMatrix());
-        setMatrix4("viewProjMatrix", camera.getViewProjectionMatrix());
-        setMatrix4("invProjMatrix", camera.getInverseProjectionMatrix());
-    }
-
-    @Override
     public boolean supportsFeature(ShaderProgramFeature feature) {
         return shader.getAvailableFeatures().contains(feature);
-    }
-
-    @Override
-    public void setFloat(String desc, float f) {
-        setFloat(desc, f, false);
     }
 
     @Override
@@ -289,11 +272,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
     }
 
     @Override
-    public void setFloat1(String desc, FloatBuffer buffer) {
-        setFloat1(desc, buffer, false);
-    }
-
-    @Override
     public void setFloat1(String desc, FloatBuffer buffer, boolean currentOnly) {
         if (isDisposed()) {
             return;
@@ -314,11 +292,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
 
             restoreStateAfterUniformsSet();
         }
-    }
-
-    @Override
-    public void setFloat2(String desc, float f1, float f2) {
-        setFloat2(desc, f1, f2, false);
     }
 
     @Override
@@ -345,21 +318,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
     }
 
     @Override
-    public void setFloat2(String name, Tuple2f value) {
-        setFloat2(name, value.x, value.y);
-    }
-
-    @Override
-    public void setFloat2(String name, Tuple2f value, boolean currentOnly) {
-        setFloat2(name, value.x, value.y, currentOnly);
-    }
-
-    @Override
-    public void setFloat2(String desc, FloatBuffer buffer) {
-        setFloat2(desc, buffer, false);
-    }
-
-    @Override
     public void setFloat2(String desc, FloatBuffer buffer, boolean currentOnly) {
         if (isDisposed()) {
             return;
@@ -380,11 +338,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
 
             restoreStateAfterUniformsSet();
         }
-    }
-
-    @Override
-    public void setFloat3(String desc, float f1, float f2, float f3) {
-        setFloat3(desc, f1, f2, f3, false);
     }
 
     @Override
@@ -411,21 +364,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
     }
 
     @Override
-    public void setFloat3(String name, Tuple3f value) {
-        setFloat3(name, value.x, value.y, value.z);
-    }
-
-    @Override
-    public void setFloat3(String name, Tuple3f value, boolean currentOnly) {
-        setFloat3(name, value.x, value.y, value.z, currentOnly);
-    }
-
-    @Override
-    public void setFloat3(String desc, FloatBuffer buffer) {
-        setFloat3(desc, buffer, false);
-    }
-
-    @Override
     public void setFloat3(String desc, FloatBuffer buffer, boolean currentOnly) {
         if (isDisposed()) {
             return;
@@ -446,11 +384,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
 
             restoreStateAfterUniformsSet();
         }
-    }
-
-    @Override
-    public void setFloat4(String desc, float f1, float f2, float f3, float f4) {
-        setFloat4(desc, f1, f2, f3, f4, false);
     }
 
     @Override
@@ -477,21 +410,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
     }
 
     @Override
-    public void setFloat4(String name, Tuple4f value) {
-        setFloat4(name, value.x, value.y, value.z, value.w);
-    }
-
-    @Override
-    public void setFloat4(String name, Tuple4f value, boolean currentOnly) {
-        setFloat4(name, value.x, value.y, value.z, value.w, currentOnly);
-    }
-
-    @Override
-    public void setFloat4(String desc, FloatBuffer buffer) {
-        setFloat4(desc, buffer, false);
-    }
-
-    @Override
     public void setFloat4(String desc, FloatBuffer buffer, boolean currentOnly) {
         if (isDisposed()) {
             return;
@@ -512,11 +430,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
 
             restoreStateAfterUniformsSet();
         }
-    }
-
-    @Override
-    public void setInt(String desc, int i) {
-        setInt(desc, i, false);
     }
 
     @Override
@@ -543,11 +456,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
     }
 
     @Override
-    public void setBoolean(String desc, boolean value) {
-        setBoolean(desc, value, false);
-    }
-
-    @Override
     public void setBoolean(String desc, boolean value, boolean currentOnly) {
         if (isDisposed()) {
             return;
@@ -568,11 +476,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
 
             restoreStateAfterUniformsSet();
         }
-    }
-
-    @Override
-    public void setMatrix3(String desc, Matrix3f value) {
-        setMatrix3(desc, value, false);
     }
 
     @Override
@@ -599,11 +502,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
     }
 
     @Override
-    public void setMatrix3(String desc, FloatBuffer value) {
-        setMatrix3(desc, value, false);
-    }
-
-    @Override
     public void setMatrix3(String desc, FloatBuffer value, boolean currentOnly) {
         if (isDisposed()) {
             return;
@@ -627,11 +525,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
     }
 
     @Override
-    public void setMatrix4(String desc, Matrix4f value) {
-        setMatrix4(desc, value, false);
-    }
-
-    @Override
     public void setMatrix4(String desc, Matrix4f value, boolean currentOnly) {
         if (isDisposed()) {
             return;
@@ -652,11 +545,6 @@ public class GLSLMaterial extends AbstractAsset<MaterialData> implements Materia
 
             restoreStateAfterUniformsSet();
         }
-    }
-
-    @Override
-    public void setMatrix4(String desc, FloatBuffer value) {
-        setMatrix3(desc, value, false);
     }
 
     @Override
